Task.Func -> Task.Handler
Func is a little confusing to read in this context and could be
easily mistaken for an anonymous function definition at first
glance.

In addition you already have this pattern in the sibling commands
package, and there the name is Handler - it's best to pick and
stick to a convention per project.